+++ kinto_http/tests/test_session.py
@@ -40,8 +40,10 @@ class SessionTest(unittest.TestCase):
         self.assertEqual(session.auth, None)
         session.request("get", "/test")
         self.requests_mock.request.assert_called_with(
-            "get", "https://example.org/test", timeout=None,
-            headers=self.requests_mock.request.headers
+            "get",
+            "https://example.org/test",
+            timeout=None,
+            headers=self.requests_mock.request.headers,
         )
 
     def test_timeout_can_be_set_to_value(self):
@@ -51,8 +53,10 @@ class SessionTest(unittest.TestCase):
         self.assertEqual(session.auth, None)
         session.request("get", "/test")
         self.requests_mock.request.assert_called_with(
-            "get", "https://example.org/test", timeout=4,
-            headers=self.requests_mock.request.headers
+            "get",
+            "https://example.org/test",
+            timeout=4,
+            headers=self.requests_mock.request.headers,
         )
 
     def test_no_auth_is_used_by_default(self):